chore : 3차 피드백 수용
